Change width and height to u32
I changed width and height to be u32 instead u16 because I noticed the
values after always seemed to be zero. This might not be correct but the
only way to determine it would be to inspect the assembly code or use a
debugger and watch the memory address.

+++ src/main.rs
@@ -32,10 +32,10 @@ struct Opt {
     /// Screen resolution height
     // The 'h' is already used by the help parameter.
     #[structopt(short = "y", long = "height", default_value = "600")]
-    height: u16,
+    height: u32,
     /// Screen resolution width
     #[structopt(short = "x", long = "width", default_value = "800")]
-    width: u16,
+    width: u32,
 }
 
 fn main() -> std::io::Result<()> {
